Ritik's age comparison (lecture 23)
Rithik wants to write a program to compare the ages of individuals. Help him to implement a program that defines a Person class to represent individuals by their ages. The program should take user input for the ages of three people and determine if Person 1 is of the same age as Person 2 or not and if Person 1 is of the same age as Person 3 or not.